+++ NodeBindings/CHANGELOG.md
@@ -1,3 +1,6 @@
+[25/04/2017]
+- Retry TakePhoto when it fails with CaptureError state change
+
 [24/04/2017]
 - 1300D TakePhoto works properly when live view is started before
 - Live view is not stopped before taking photo and started again after